[#23] [ADD] hightlightForQuizService 추가 (하이라이트 관련 서비스 분리)
